Reduce warning

+++ ReFlanker/NSData+MD5Digest.m
@@ -14,7 +14,7 @@
 - (NSString *)MD5Digest
 {
     unsigned char md5Buffer[CC_MD5_DIGEST_LENGTH];
-    CC_MD5(self.bytes, self.length, md5Buffer);
+    CC_MD5(self.bytes, (unsigned int)self.length, md5Buffer);
     
     NSMutableString *output = [NSMutableString stringWithCapacity:CC_MD5_DIGEST_LENGTH * 2];
     for(int i = 0; i < CC_MD5_DIGEST_LENGTH; i++) {
